+++ pkg/scheduler/api/podgroup_info/subgroup_info/factory.go
@@ -74,7 +74,7 @@ func createSubGroupInfos(allSubGroups map[string]*v2alpha2.SubGroup, children ma
 		if hasChildren {
 			subGroupSets[name] = NewSubGroupSet(name, topologyConstrainInfo)
 		} else {
-			podSets[name] = NewPodSet(name, subGroup.MinMember, topologyConstrainInfo)
+			podSets[name] = NewPodSet(name, max(subGroup.MinMember, 1), topologyConstrainInfo)
 		}
 	}
 }